Revert "Delete VisualStudioMetadataReference, which was now an empty type"
This reverts commit 73f1152e961c57e68c34331040e94b60c14ead24. Right now
F# is still implicitly mentioning this type. This has been fixed in
https://github.com/Microsoft/visualfsharp/pull/6153 but until we can
get that rolled out we'll roll back the name change.
